Added more complex authoriation.

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -1,12 +1,16 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Models;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
+using System.Threading.Tasks;
 
 namespace RestaurantAPI.Services
 {
@@ -15,12 +19,17 @@ namespace RestaurantAPI.Services
         private readonly RestaurantDbContext _context;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService _userContextService;
 
-        public RestaurantService(RestaurantDbContext context, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext context, IMapper mapper, ILogger<RestaurantService> logger, IAuthorizationService authorizationService,
+            IUserContextService userContextService)
         {
             _context = context;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
+            _userContextService = userContextService;
         }
         public RestaurantDto GetById(int id)
         {
@@ -37,13 +46,13 @@ namespace RestaurantAPI.Services
             return resturantDto;
         }
 
-        public IEnumerable<RestaurantDto> GetAll()
+        public async Task<IEnumerable<RestaurantDto>> GetAll()
         {
-            var resturants = _context
+            var resturants = await _context
               .Restaurants
               .Include(r => r.Address)
               .Include(r => r.Dishes)
-              .ToList();
+              .ToListAsync();
 
             var resturantsDtos = _mapper.Map<List<RestaurantDto>>(resturants);
 
@@ -53,6 +62,7 @@ namespace RestaurantAPI.Services
         public int CreateRestaurant(CreateRestaurantDto dto)
         {
             var newRestaurant = _mapper.Map<Restaurant>(dto);
+            newRestaurant.CreatedById = _userContextService.GetUserId;
 
             _context.Restaurants.Add(newRestaurant);
             _context.SaveChanges();
@@ -67,6 +77,10 @@ namespace RestaurantAPI.Services
 
             if(restaurantToDelete == null) throw new NotFoundException("Restaurant not found");
 
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurantToDelete, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded) throw new ForbidException();
+
             _context.Restaurants.Remove(restaurantToDelete);
             _context.SaveChanges();
         }
@@ -76,6 +90,10 @@ namespace RestaurantAPI.Services
             var restaurant = _context.Restaurants.SingleOrDefault(x => x.Id == id);
 
             if (restaurant == null) throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded) throw new ForbidException();
 
             restaurant.Name = updateRestaurantDto.Name;
             restaurant.Description = updateRestaurantDto.Description;
